src: Remove SmartComponents/PresentationalComponents split
Just have a directory per component.
