+++ src/Components/CreateImageWizard/CreateImageWizard.scss
