+++ src/Components/CreateImageWizard/ImageWizardFooter.js
