+++ src/Components/CreateImageWizard/ImageWizardFooter.scss
