+++ src/Components/CreateImageWizard/WizardStepImageOutput.scss
