+++ src/Components/CreateImageWizard/WizardStepRegistration.js
