+++ src/Components/CreateImageWizard/WizardStepReview.scss
