+++ src/Components/CreateImageWizard/WizardStepUploadAzure.js
