+++ src/Components/CreateImageWizard/WizardStepUploadAzure.scss
