+++ src/test/redux/reducers.test.js
@@ -1,5 +1,5 @@
-import { composes } from '../../../store/reducers/composes';
-import types from '../../../store/types';
+import { composes } from '../../store/reducers/composes';
+import types from '../../store/types';
 
 const compose = {
     id: '77e4c693-0497-4b85-936d-b2a3ad69571b',
